Adds typical location of Postgresql used by One Click Installer, and reminder to wrap paths containing spaces in quotes
The Postgres path typically used by the One Click Installer contains spaces. If users forget to wrap the path in quotes then cl.exe will fail, as it will not find the correct path. This may be obvious to experienced users but beginners or intermediate users such as myself need reminding.